RAG-Vector DB application/requirements.txt
numpy
pandas
langchain
langchain_community
langchain-openai
langchain_core
openai
langchain_huggingface
fastapi
ipykernel
uvicorn
python-dotenv
pymongo
pypdf2